app/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
fastapi==0.115.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.1
pocketbase==0.12.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.1
